Move subfunctions into master location tools file
